refactor(multitask): extract launcher, state, utils modules from orchestrator.rs
Split orchestrator.rs (1135 LOC) into 4 focused modules:
- orchestrator.rs (400 LOC): core orchestration loop, monitor, report
- launcher.rs (256 LOC): launcher generation, process spawning, delegate detection
- state.rs (338 LOC): state persistence, lifecycle management, cleanup
- utils.rs (161 LOC): atomic_write, tail_output, disk space, dir_size

Moved tests to co-locate with their tested code (models.rs, launcher.rs,
state.rs, utils.rs). Pure structural refactoring — no behavior changes.
