+++ crates/amplihack-cli/src/commands/multitask/models.rs
@@ -209,3 +209,79 @@ pub struct PersistedState {
     #[serde(default)]
     pub resume_context: Option<HashMap<String, serde_json::Value>>,
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_sanitize_id() {
+        assert_eq!(sanitize_id("123"), "123");
+        assert_eq!(sanitize_id("hello-world"), "hello-world");
+        assert_eq!(sanitize_id("path/../../etc"), "path_______etc");
+        assert_eq!(sanitize_id("a b c"), "a_b_c");
+    }
+
+    #[test]
+    fn test_workstream_config_parsing() {
+        let json = r#"[
+            {
+                "issue": 42,
+                "branch": "feat/test",
+                "description": "Test workstream",
+                "task": "Do something",
+                "recipe": "default-workflow"
+            }
+        ]"#;
+        let configs: Vec<WorkstreamConfig> = serde_json::from_str(json).unwrap();
+        assert_eq!(configs.len(), 1);
+        assert_eq!(configs[0].issue_id(), 42);
+        assert_eq!(configs[0].branch, "feat/test");
+        assert_eq!(configs[0].description_or_default(), "Test workstream");
+    }
+
+    #[test]
+    fn test_workstream_config_string_issue() {
+        let json = r#"[{"issue": "99", "branch": "b", "task": "t"}]"#;
+        let configs: Vec<WorkstreamConfig> = serde_json::from_str(json).unwrap();
+        assert_eq!(configs[0].issue_id(), 99);
+    }
+
+    #[test]
+    fn test_workstream_config_default_description() {
+        let json = r#"[{"issue": 7, "branch": "b", "task": "t"}]"#;
+        let configs: Vec<WorkstreamConfig> = serde_json::from_str(json).unwrap();
+        assert_eq!(configs[0].description_or_default(), "Issue #7");
+    }
+
+    #[test]
+    fn test_persisted_state_roundtrip() {
+        let state = PersistedState {
+            issue: 42,
+            branch: "feat/test".to_string(),
+            lifecycle_state: "completed".to_string(),
+            ..Default::default()
+        };
+        let json = serde_json::to_string(&state).unwrap();
+        let parsed: PersistedState = serde_json::from_str(&json).unwrap();
+        assert_eq!(parsed.issue, 42);
+        assert_eq!(parsed.branch, "feat/test");
+        assert_eq!(parsed.lifecycle_state, "completed");
+    }
+
+    #[test]
+    fn test_derive_cleanup_eligible() {
+        assert!(Workstream::derive_cleanup_eligible("completed"));
+        assert!(Workstream::derive_cleanup_eligible("failed_terminal"));
+        assert!(Workstream::derive_cleanup_eligible("abandoned"));
+        assert!(!Workstream::derive_cleanup_eligible("running"));
+        assert!(!Workstream::derive_cleanup_eligible("pending"));
+        assert!(!Workstream::derive_cleanup_eligible("failed_resumable"));
+    }
+
+    #[test]
+    fn test_default_constants() {
+        assert_eq!(DEFAULT_MAX_RUNTIME, 7200);
+        assert_eq!(DEFAULT_TIMEOUT_POLICY, "interrupt-preserve");
+    }
+}
